Upgrade major deps: MUI 6->9, Vite 5->8, wagmi 2->3
