+++ package.json
@@ -6,7 +6,7 @@
   "private": false,
   "type": "module",
   "engines": {
-    "node": ">=20.0.0"
+    "node": ">=20.19.0"
   },
   "scripts": {
     "dev": "vite",
@@ -21,13 +21,13 @@
   "dependencies": {
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
-    "@mui/icons-material": "^6.1.6",
-    "@mui/material": "^6.1.6",
-    "@tanstack/react-query": "^5.59.20",
+    "@mui/icons-material": "^9.0.1",
+    "@mui/material": "^9.0.1",
+    "@tanstack/react-query": "^5.100.14",
     "aos": "^2.3.4",
-    "jarallax": "^2.2.0",
+    "jarallax": "^3.0.0",
     "numeral": "^2.0.6",
-    "octokit": "^4.0.2",
+    "octokit": "^5.0.5",
     "prop-types": "^15.8.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -35,19 +35,19 @@
     "react-router-dom": "^7.0.1",
     "react-type-animation": "^3.2.0",
     "sass": "^1.80.6",
-    "sonner": "^1.7.4",
+    "sonner": "^2.0.7",
     "use-local-storage-state": "^19.5.0",
-    "viem": "^2.21.42",
-    "wagmi": "^2.12.31",
-    "web-vitals": "^4.2.4"
+    "viem": "^2.51.0",
+    "wagmi": "^3.6.15",
+    "web-vitals": "^5.2.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-react": "^4.3.3",
+    "@vitejs/plugin-react": "^6.0.2",
     "eslint": "^9.14.0",
     "eslint-plugin-react": "^7.37.2",
     "eslint-plugin-react-hooks": "^5.0.0",
     "prettier": "^3.3.3",
-    "vite": "^5.4.10",
+    "vite": "^8.0.14",
     "wrangler": "^4.0.0"
   }
 }
