+++ src/views/Staking/components/Hero/Hero.jsx
