+++ src/views/Vote/Vote.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Main from 'layouts/Main';
+import Container from 'components/Container';
+import { Functions } from './components';
+
+const Vote = () => {
+  return (
+    <Main>
+      <Container>
+        <Functions />
+      </Container>
+    </Main>
+  );
+};
+
+export default Vote;
