+++ src/views/Vote/components/Functions/Functions.jsx
