+++ src/views/Vote/components/Proposals/Proposals.jsx
